feat: Add average rating display to laptop list and detail pages
This commit adds the display of average ratings to the laptop list and detail pages. It includes the modification of the laptop_list.html and laptop_detail.html templates to show the average rating of each laptop. This change provides users with valuable information about the ratings of the laptops.